+++ aws-java-sdk-s3/src/main/java/com/amazonaws/services/s3/model/SetBucketReplicationConfigurationRequest.java
@@ -33,6 +33,8 @@ public class SetBucketReplicationConfigurationRequest extends
      */
     private BucketReplicationConfiguration replicationConfiguration;
 
+    private String token;
+
     /**
      * Creates a new SetReplicationConfigurationRequest.
      */
@@ -120,6 +122,19 @@ public class SetBucketReplicationConfigurationRequest extends
             BucketReplicationConfiguration replicationConfiguration) {
         setReplicationConfiguration(replicationConfiguration);
         return this;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public SetBucketReplicationConfigurationRequest withToken(String token) {
+        setToken(token);
+        return this;
+    }
+
+    public String getToken() {
+        return token;
     }
 
     @Override